+++ CHANGELOG.md
@@ -20,3 +20,8 @@
 * Replaced references to the former `GenTreeSrc` class in `README.md` with `Main`.
 - Made the `Main` class uninstantiable by making it `final` and making the default constructor `private`.
 - Removed the `PropertyStyle` class.
+
+# GenTreeSrc v0.3.0
+
++ Add a `javax.tools.Tool` implementation.
+* Updated command-line parsing with new `-o` option for output directory.